Problem
Aria-live of TreeSelect option was always updated with the value of the option.
In some cases there is some sort of ID in the value field (not relevant for user),
and the textual representation in label. In those cases, users
with screenreaders are confused, as they hear only some ID-numbers instead
of option labels. This is a huge problem for accessibility.

Solution

Allow developer to decide whether to use option label in aria-live area.
There is a new prop ´bool labelInAriaLive (default=false)´ , that
the developer can use if he/she wishes. This is a huge enabler for
improving accessibility.

Notes

Also testcase was added and documentation was updated in this commit.
This commit will not break any existing usage of treeSelect component,
as the default for new prop is false. If ´labelInAriaLive´ is false,
everything will work as expected.